Create Chapter 5 #2
Change int to double on line 9.
Add a semi colon at the end of line 10.
Change "return int" to return k.
shance cin>>d to cin>>c.
c should not have quotation marks around it.
Cout should have a lower case c.

Not the most efficient way of doing it. But it is what it is.